ref(tracing): Move `extractTraceparentData` into `@sentry/utils`
