+++ packages/utils/src/tracing.ts
@@ -1,0 +1,34 @@
+import { TraceparentData } from '@sentry/types';
+
+export const TRACEPARENT_REGEXP = new RegExp(
+  '^[ \\t]*' + // whitespace
+    '([0-9a-f]{32})?' + // trace_id
+    '-?([0-9a-f]{16})?' + // span_id
+    '-?([01])?' + // sampled
+    '[ \\t]*$', // whitespace
+);
+
+/**
+ * Extract transaction context data from a `sentry-trace` header.
+ *
+ * @param traceparent Traceparent string
+ *
+ * @returns Object containing data from the header, or undefined if traceparent string is malformed
+ */
+export function extractTraceparentData(traceparent: string): TraceparentData | undefined {
+  const matches = traceparent.match(TRACEPARENT_REGEXP);
+  if (matches) {
+    let parentSampled: boolean | undefined;
+    if (matches[3] === '1') {
+      parentSampled = true;
+    } else if (matches[3] === '0') {
+      parentSampled = false;
+    }
+    return {
+      traceId: matches[1],
+      parentSampled,
+      parentSpanId: matches[2],
+    };
+  }
+  return undefined;
+}
